MobihiveLabs/requirements.txt
python==3.10.1
Django==4.0.4
gunicorn==20.1.0
Pillow==9.1.0
social-auth-app-django==3.1.0
django-extensions==3.6.0
# django-extensions
werkzeug==2.2.2
pyOpenSSL==22.0.0
qrcode==7.3.1
django-qr-code